requirements.txt
mypy
flake8
black
types-requests
pillow
pyyaml